Removed the security warning for a search within the site
After clicking `<Enter>` in the search box in Firefox, there will be a
popup dialog box which says "Although this page is encrypted, the
information you have entered is to be sent over an unencrypted
connection...".  That's quite annoying.  Therefore, I added an `s` to
the URL of the search engine.