Corrected the mistake in my previous commit
Commit 449a911 did remove the repeating images in category feed pages.
However, it disabled the use of image popup plugin in the blog posts
appeared in those category feed pages.  Now, this bug should have been
fixed.

I'm writing a post, and have forgotten to move them out of my working
Octopress folder before generating the site and deploying it to GitHub.
As a result, part of it is leaked out, and the new post is TBC.